Android Drop
It's interesting that Kaspersky flags it as an iPhoneOS trojan. According to VirusTotal, it's actually an APK file, yet multiple antivirus engines are identifying it as a MacOS or iOS threat. Some AVs do detect it as a trojan for Android.

Hate when it has a HEUR in it which is from a different OS. Is it iOS, is it Android?